Rename UnitOfWorkPipelineBehaviorConcurrentCounter to ConcurrentCommandCounter, delay telemetry events until all commands are processed
